chore: update post service communication with API Gateway

+++ APIGateway/app/Http/Controllers/PostController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Contracts\PostServiceInterface;
-use App\Contracts\UserServiceInterface;
+use App\Contracts\AuthServiceInterface;
 use App\Traits\ApiResponseTrait;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
@@ -16,24 +16,22 @@ class PostController extends Controller
 
     public PostServiceInterface $postService;
 
-    public UserServiceInterface $userService;
+    public AuthServiceInterface $userService;
 
     /**
      * @param PostServiceInterface $postService
-     * @param UserServiceInterface $userService
      */
-    public function __construct(PostServiceInterface $postService, UserServiceInterface $userService)
+    public function __construct(PostServiceInterface $postService)
     {
         $this->postService = $postService;
-        $this->userService = $userService;
     }
 
     /**
      * Return list of posts
      *
-     * @return Response|ResponseFactory
+     * @return JsonResponse
      */
-    public function index(): Response|ResponseFactory
+    public function index(): JsonResponse
     {
         return $this->success($this->postService->index());
     }
@@ -42,24 +40,28 @@ class PostController extends Controller
      * Create new post
      *
      * @param Request $request
-     * @return Response|ResponseFactory
+     * @return JsonResponse
      */
-    public function store(Request $request): Response|ResponseFactory
+    public function store(Request $request): JsonResponse
     {
-        $this->userService->show($request->user_id);
+        $payload = $this->attachUserToPayload($request);
 
-        return $this->success($this->postService->store($request->all()));
+        $response = $this->postService->store($payload);
+
+        return $this->success($response, 'Post successfully created', Response::HTTP_CREATED);
     }
 
     /**
      * get and show details of existing post
      *
      * @param $post
-     * @return Response|ResponseFactory
+     * @return JsonResponse
      */
-    public function show($post): Response|ResponseFactory
+    public function show($post): JsonResponse
     {
-        return $this->success($this->postService->show($post));
+        $response = $this->postService->show($post);
+
+        return $this->success($response);
     }
 
     /**
@@ -67,13 +69,15 @@ class PostController extends Controller
      *
      * @param Request $request
      * @param $post
-     * @return Response|ResponseFactory
+     * @return JsonResponse
      */
-    public function update(Request $request, $post): Response|ResponseFactory
+    public function update(Request $request, $post): JsonResponse
     {
-        $this->userService->show($request->user_id);
+        $payload = $this->attachUserToPayload($request);
 
-        return $this->success($this->postService->update($request->all(), $post));
+        $response = $this->postService->update($payload, $post);
+
+        return $this->success($response, 'Post successfully updated');
     }
 
     /**
@@ -81,13 +85,15 @@ class PostController extends Controller
      *
      * @param Request $request
      * @param int $post
-     * @return Response|ResponseFactory
+     * @return JsonResponse
      */
-    public function like(Request $request, int $post): Response|ResponseFactory
+    public function like(Request $request, int $post): JsonResponse
     {
-        $this->userService->show($request->user_id);
+        $payload = $this->attachUserToPayload($request);
 
-        return $this->success($this->postService->like($request->all(), $post));
+        $response = $this->postService->like($payload, $post);
+
+        return $this->success($response, 'Post liked');
     }
 
     /**
@@ -98,8 +104,22 @@ class PostController extends Controller
      */
     public function destroy($post): JsonResponse
     {
+        $getPost = json_decode($this->postService->show($post));
+
+        if ($getPost->user_id !== auth()->user()->id) {
+            return $this->error('You are not authorized to delete this post', Response::HTTP_FORBIDDEN);
+        }
+
         $this->postService->delete($post);
 
         return $this->deleted();
+    }
+
+    protected function attachUserToPayload(Request $request): array {
+        $userId = auth()->user()->id;
+        $payload = $request->all();
+        $payload['user_id'] = $userId;
+
+        return $payload;
     }
 }
